requirements.txt
python-telegram-bot==20.8
apscheduler==3.10.4
aiohttp==3.9.5
requests==2.32.3
nest_asyncio==1.6.0
